requirements.txt
lightrag-hku
google-generativeai
pypdf
